ライブラリがバグってると困るんだよなぁ

+++ prime_factorization.hpp
@@ -3,7 +3,7 @@
 
 map<ll, ll> prime_factorization(ll n) {
   map<ll, ll> m;
-  for (int i = 2; i*i <= n; i++) {
+  for (ll i = 2; i*i <= n; i++) {
     while (n%i == 0) {
       m[i]++;
       n /= i;
